# add benchmark for sql.RawBytes