@methane 2019-04-03
after.txt is #943.

```
$ benchcmp before.txt after.txt
benchmark                              old ns/op     new ns/op     delta
BenchmarkQueryRawBytes/size=100        2082425       2065162       -0.83%
BenchmarkQueryRawBytes/size=1000       2114160       2101272       -0.61%
BenchmarkQueryRawBytes/size=2000       2131505       2135560       +0.19%
BenchmarkQueryRawBytes/size=4000       2150739       2146939       -0.18%
BenchmarkQueryRawBytes/size=8000       2195996       2190682       -0.24%
BenchmarkQueryRawBytes/size=12000      2290469       2256546       -1.48%
BenchmarkQueryRawBytes/size=16000      2371228       2329660       -1.75%
BenchmarkQueryRawBytes/size=32000      2705139       2801268       +3.55%
BenchmarkQueryRawBytes/size=64000      3353764       3495100       +4.21%
BenchmarkQueryRawBytes/size=256000     6925304       7302883       +5.45%

benchmark                              old allocs     new allocs     delta
BenchmarkQueryRawBytes/size=100        223            223            +0.00%
BenchmarkQueryRawBytes/size=1000       223            223            +0.00%
BenchmarkQueryRawBytes/size=2000       223            223            +0.00%
BenchmarkQueryRawBytes/size=4000       223            223            +0.00%
BenchmarkQueryRawBytes/size=8000       223            223            +0.00%
BenchmarkQueryRawBytes/size=12000      223            223            +0.00%
BenchmarkQueryRawBytes/size=16000      223            223            +0.00%
BenchmarkQueryRawBytes/size=32000      223            323            +44.84%
BenchmarkQueryRawBytes/size=64000      223            323            +44.84%
BenchmarkQueryRawBytes/size=256000     223            323            +44.84%

benchmark                              old bytes     new bytes     delta
BenchmarkQueryRawBytes/size=100        7072          7072          +0.00%
BenchmarkQueryRawBytes/size=1000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=2000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=4000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=8000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=12000      7072          7072          +0.00%
BenchmarkQueryRawBytes/size=16000      7072          7072          +0.00%
BenchmarkQueryRawBytes/size=32000      7072          3283877       +46334.91%
BenchmarkQueryRawBytes/size=64000      7072          6560680       +92669.80%
BenchmarkQueryRawBytes/size=256000     7072          26221480      +370678.85%
```