@methane 2019-04-03
pr905.txt is #905.
```
$ benchcmp before.txt pr905.txt
benchmark                              old ns/op     new ns/op     delta
BenchmarkQueryRawBytes/size=100        2082425       2042182       -1.93%
BenchmarkQueryRawBytes/size=1000       2114160       2056612       -2.72%
BenchmarkQueryRawBytes/size=2000       2131505       2075042       -2.65%
BenchmarkQueryRawBytes/size=4000       2150739       2100935       -2.32%
BenchmarkQueryRawBytes/size=8000       2195996       2168976       -1.23%
BenchmarkQueryRawBytes/size=12000      2290469       2248544       -1.83%
BenchmarkQueryRawBytes/size=16000      2371228       2297753       -3.10%
BenchmarkQueryRawBytes/size=32000      2705139       2681204       -0.88%
BenchmarkQueryRawBytes/size=64000      3353764       3257480       -2.87%
BenchmarkQueryRawBytes/size=256000     6925304       6912481       -0.19%

benchmark                              old allocs     new allocs     delta
BenchmarkQueryRawBytes/size=100        223            218            -2.24%
BenchmarkQueryRawBytes/size=1000       223            218            -2.24%
BenchmarkQueryRawBytes/size=2000       223            218            -2.24%
BenchmarkQueryRawBytes/size=4000       223            218            -2.24%
BenchmarkQueryRawBytes/size=8000       223            218            -2.24%
BenchmarkQueryRawBytes/size=12000      223            218            -2.24%
BenchmarkQueryRawBytes/size=16000      223            218            -2.24%
BenchmarkQueryRawBytes/size=32000      223            218            -2.24%
BenchmarkQueryRawBytes/size=64000      223            218            -2.24%
BenchmarkQueryRawBytes/size=256000     223            218            -2.24%

benchmark                              old bytes     new bytes     delta
BenchmarkQueryRawBytes/size=100        7072          7000          -1.02%
BenchmarkQueryRawBytes/size=1000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=2000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=4000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=8000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=12000      7072          7000          -1.02%
BenchmarkQueryRawBytes/size=16000      7072          7000          -1.02%
BenchmarkQueryRawBytes/size=32000      7072          7000          -1.02%
BenchmarkQueryRawBytes/size=64000      7072          7000          -1.02%
BenchmarkQueryRawBytes/size=256000     7072          7000          -1.02%
```