@methane 2019-04-03
#943 vs #905 

```
$ benchcmp after.txt pr905.txt
benchmark                              old ns/op     new ns/op     delta
BenchmarkQueryRawBytes/size=100        2065162       2042182       -1.11%
BenchmarkQueryRawBytes/size=1000       2101272       2056612       -2.13%
BenchmarkQueryRawBytes/size=2000       2135560       2075042       -2.83%
BenchmarkQueryRawBytes/size=4000       2146939       2100935       -2.14%
BenchmarkQueryRawBytes/size=8000       2190682       2168976       -0.99%
BenchmarkQueryRawBytes/size=12000      2256546       2248544       -0.35%
BenchmarkQueryRawBytes/size=16000      2329660       2297753       -1.37%
BenchmarkQueryRawBytes/size=32000      2801268       2681204       -4.29%
BenchmarkQueryRawBytes/size=64000      3495100       3257480       -6.80%
BenchmarkQueryRawBytes/size=256000     7302883       6912481       -5.35%

benchmark                              old allocs     new allocs     delta
BenchmarkQueryRawBytes/size=100        223            218            -2.24%
BenchmarkQueryRawBytes/size=1000       223            218            -2.24%
BenchmarkQueryRawBytes/size=2000       223            218            -2.24%
BenchmarkQueryRawBytes/size=4000       223            218            -2.24%
BenchmarkQueryRawBytes/size=8000       223            218            -2.24%
BenchmarkQueryRawBytes/size=12000      223            218            -2.24%
BenchmarkQueryRawBytes/size=16000      223            218            -2.24%
BenchmarkQueryRawBytes/size=32000      323            218            -32.51%
BenchmarkQueryRawBytes/size=64000      323            218            -32.51%
BenchmarkQueryRawBytes/size=256000     323            218            -32.51%

benchmark                              old bytes     new bytes     delta
BenchmarkQueryRawBytes/size=100        7072          7000          -1.02%
BenchmarkQueryRawBytes/size=1000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=2000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=4000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=8000       7072          7000          -1.02%
BenchmarkQueryRawBytes/size=12000      7072          7000          -1.02%
BenchmarkQueryRawBytes/size=16000      7072          7000          -1.02%
BenchmarkQueryRawBytes/size=32000      3283877       7000          -99.79%
BenchmarkQueryRawBytes/size=64000      6560680       7000          -99.89%
BenchmarkQueryRawBytes/size=256000     26221480      7000          -99.97%
```