@methane 2019-04-03
Hmm, both PRs affects to all queries, not only RawBytes.
This is benchcmp for all benchmarks.

```
$ benchcmp pr905all.txt pr943all.txt
benchmark                              old ns/op     new ns/op     delta
BenchmarkQuery                         17160         18327         +6.80%
BenchmarkExec                          10472         11837         +13.03%
BenchmarkRoundtripTxt                  100658        98245         -2.40%
BenchmarkRoundtripBin                  73582         70721         -3.89%
BenchmarkInterpolation                 643           637           -0.93%
BenchmarkQueryContext/1                48208         56076         +16.32%
BenchmarkQueryContext/2                26755         25814         -3.52%
BenchmarkQueryContext/3                18533         19206         +3.63%
BenchmarkQueryContext/4                14533         17886         +23.07%
BenchmarkExecContext/1                 50131         50246         +0.23%
BenchmarkExecContext/2                 26080         28813         +10.48%
BenchmarkExecContext/3                 18580         19151         +3.07%
BenchmarkExecContext/4                 16396         17596         +7.32%
BenchmarkQueryRawBytes/size=100        2112559       2085116       -1.30%
BenchmarkQueryRawBytes/size=1000       2127677       2138447       +0.51%
BenchmarkQueryRawBytes/size=2000       2149052       2175069       +1.21%
BenchmarkQueryRawBytes/size=4000       2176124       2185810       +0.45%
BenchmarkQueryRawBytes/size=8000       2233973       2215323       -0.83%
BenchmarkQueryRawBytes/size=12000      2329292       2294619       -1.49%
BenchmarkQueryRawBytes/size=16000      2462832       2357734       -4.27%
BenchmarkQueryRawBytes/size=32000      2677587       2905703       +8.52%
BenchmarkQueryRawBytes/size=64000      3529586       3521465       -0.23%
BenchmarkQueryRawBytes/size=256000     7132580       7905538       +10.84%
BenchmarkParseDSN                      6911          6981          +1.01%

benchmark                              old allocs     new allocs     delta
BenchmarkQuery                         18             23             +27.78%
BenchmarkExec                          3              8              +166.67%
BenchmarkRoundtripTxt                  12             17             +41.67%
BenchmarkRoundtripBin                  14             18             +28.57%
BenchmarkInterpolation                 1              1              +0.00%
BenchmarkQueryContext/1                17             22             +29.41%
BenchmarkQueryContext/2                17             22             +29.41%
BenchmarkQueryContext/3                17             22             +29.41%
BenchmarkQueryContext/4                17             22             +29.41%
BenchmarkExecContext/1                 17             22             +29.41%
BenchmarkExecContext/2                 17             22             +29.41%
BenchmarkExecContext/3                 17             22             +29.41%
BenchmarkExecContext/4                 17             22             +29.41%
BenchmarkQueryRawBytes/size=100        218            223            +2.29%
BenchmarkQueryRawBytes/size=1000       218            223            +2.29%
BenchmarkQueryRawBytes/size=2000       218            223            +2.29%
BenchmarkQueryRawBytes/size=4000       218            223            +2.29%
BenchmarkQueryRawBytes/size=8000       218            223            +2.29%
BenchmarkQueryRawBytes/size=12000      218            223            +2.29%
BenchmarkQueryRawBytes/size=16000      218            223            +2.29%
BenchmarkQueryRawBytes/size=32000      218            323            +48.17%
BenchmarkQueryRawBytes/size=64000      218            323            +48.17%
BenchmarkQueryRawBytes/size=256000     218            323            +48.17%
BenchmarkParseDSN                      73             73             +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkQuery                         635           609           -4.09%
BenchmarkExec                          64            136           +112.50%
BenchmarkRoundtripTxt                  36469         25565         -29.90%
BenchmarkRoundtripBin                  38062         12333         -67.60%
BenchmarkInterpolation                 176           176           +0.00%
BenchmarkQueryContext/1                618           592           -4.21%
BenchmarkQueryContext/2                618           592           -4.21%
BenchmarkQueryContext/3                618           592           -4.21%
BenchmarkQueryContext/4                618           592           -4.21%
BenchmarkExecContext/1                 618           592           -4.21%
BenchmarkExecContext/2                 618           592           -4.21%
BenchmarkExecContext/3                 618           592           -4.21%
BenchmarkExecContext/4                 618           592           -4.21%
BenchmarkQueryRawBytes/size=100        7000          7072          +1.03%
BenchmarkQueryRawBytes/size=1000       7000          7072          +1.03%
BenchmarkQueryRawBytes/size=2000       7000          7072          +1.03%
BenchmarkQueryRawBytes/size=4000       7000          7072          +1.03%
BenchmarkQueryRawBytes/size=8000       7000          7072          +1.03%
BenchmarkQueryRawBytes/size=12000      7000          7072          +1.03%
BenchmarkQueryRawBytes/size=16000      7000          7072          +1.03%
BenchmarkQueryRawBytes/size=32000      7000          3283873       +46812.47%
BenchmarkQueryRawBytes/size=64000      7000          6560673       +93623.90%
BenchmarkQueryRawBytes/size=256000     7000          26221480      +374492.57%
BenchmarkParseDSN                      7952          7952          +0.00%
```