@methane 2019-04-03
I'm sorry, #905 have not merged master branch.  master branch merged connCheck which
increase allocations.
I rebased #905 (commit = 45ca928) now.  delta is smaller than above comment.

```
$ benchcmp pr905all.txt pr943all.txt
benchmark                              old ns/op     new ns/op     delta
BenchmarkQuery                         18388         18327         -0.33%
BenchmarkExec                          11688         11837         +1.27%
BenchmarkRoundtripTxt                  100898        98245         -2.63%
BenchmarkRoundtripBin                  73160         70721         -3.33%
BenchmarkInterpolation                 648           637           -1.70%
BenchmarkQueryContext/1                49050         56076         +14.32%
BenchmarkQueryContext/2                26132         25814         -1.22%
BenchmarkQueryContext/3                19054         19206         +0.80%
BenchmarkQueryContext/4                17305         17886         +3.36%
BenchmarkExecContext/1                 49717         50246         +1.06%
BenchmarkExecContext/2                 25031         28813         +15.11%
BenchmarkExecContext/3                 19448         19151         -1.53%
BenchmarkExecContext/4                 16240         17596         +8.35%
BenchmarkQueryRawBytes/size=100        2161196       2085116       -3.52%
BenchmarkQueryRawBytes/size=1000       2174977       2138447       -1.68%
BenchmarkQueryRawBytes/size=2000       2201240       2175069       -1.19%
BenchmarkQueryRawBytes/size=4000       2224169       2185810       -1.72%
BenchmarkQueryRawBytes/size=8000       2270772       2215323       -2.44%
BenchmarkQueryRawBytes/size=12000      2337594       2294619       -1.84%
BenchmarkQueryRawBytes/size=16000      2408917       2357734       -2.12%
BenchmarkQueryRawBytes/size=32000      2733884       2905703       +6.28%
BenchmarkQueryRawBytes/size=64000      3426760       3521465       +2.76%
BenchmarkQueryRawBytes/size=256000     7023087       7905538       +12.57%
BenchmarkParseDSN                      6827          6981          +2.26%

benchmark                              old allocs     new allocs     delta
BenchmarkQuery                         23             23             +0.00%
BenchmarkExec                          8              8              +0.00%
BenchmarkRoundtripTxt                  17             17             +0.00%
BenchmarkRoundtripBin                  19             18             -5.26%
BenchmarkInterpolation                 1              1              +0.00%
BenchmarkQueryContext/1                22             22             +0.00%
BenchmarkQueryContext/2                22             22             +0.00%
BenchmarkQueryContext/3                22             22             +0.00%
BenchmarkQueryContext/4                22             22             +0.00%
BenchmarkExecContext/1                 22             22             +0.00%
BenchmarkExecContext/2                 22             22             +0.00%
BenchmarkExecContext/3                 22             22             +0.00%
BenchmarkExecContext/4                 22             22             +0.00%
BenchmarkQueryRawBytes/size=100        223            223            +0.00%
BenchmarkQueryRawBytes/size=1000       223            223            +0.00%
BenchmarkQueryRawBytes/size=2000       223            223            +0.00%
BenchmarkQueryRawBytes/size=4000       223            223            +0.00%
BenchmarkQueryRawBytes/size=8000       223            223            +0.00%
BenchmarkQueryRawBytes/size=12000      223            223            +0.00%
BenchmarkQueryRawBytes/size=16000      223            223            +0.00%
BenchmarkQueryRawBytes/size=32000      223            323            +44.84%
BenchmarkQueryRawBytes/size=64000      223            323            +44.84%
BenchmarkQueryRawBytes/size=256000     223            323            +44.84%
BenchmarkParseDSN                      73             73             +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkQuery                         699           609           -12.88%
BenchmarkExec                          136           136           +0.00%
BenchmarkRoundtripTxt                  36541         25565         -30.04%
BenchmarkRoundtripBin                  38135         12333         -67.66%
BenchmarkInterpolation                 176           176           +0.00%
BenchmarkQueryContext/1                682           592           -13.20%
BenchmarkQueryContext/2                682           592           -13.20%
BenchmarkQueryContext/3                682           592           -13.20%
BenchmarkQueryContext/4                682           592           -13.20%
BenchmarkExecContext/1                 682           592           -13.20%
BenchmarkExecContext/2                 682           592           -13.20%
BenchmarkExecContext/3                 682           592           -13.20%
BenchmarkExecContext/4                 682           592           -13.20%
BenchmarkQueryRawBytes/size=100        7072          7072          +0.00%
BenchmarkQueryRawBytes/size=1000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=2000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=4000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=8000       7072          7072          +0.00%
BenchmarkQueryRawBytes/size=12000      7072          7072          +0.00%
BenchmarkQueryRawBytes/size=16000      7072          7072          +0.00%
BenchmarkQueryRawBytes/size=32000      7072          3283873       +46334.86%
BenchmarkQueryRawBytes/size=64000      7072          6560673       +92669.70%
BenchmarkQueryRawBytes/size=256000     7072          26221480      +370678.85%
BenchmarkParseDSN                      7952          7952          +0.00%
```